ProjectManagement/requirements.txt
pandas==2.2.3
openai==1.78.1
python-dotenv==1.1.0
tenacity==9.1.4
numpy==2.4.4
anthropic==0.89.0